test: consolidate tiny component and validation tests into domain siblings
